requirements.txt
transformers
torch
fastapi
uvicorn
pillow
opencv-python
numpy
torchvision
pillow